Revert "fix: TxDataZero"
This reverts commit 2cff1787ea8d67276a36a82c8a5b6f153c435cb6.

+++ erigon-lib/rlphacks/bytes.go
@@ -34,7 +34,7 @@ func (b RlpSerializableBytes) DoubleRLPLen() int {
 	if len(b) < 1 {
 		return 0
 	}
-	return generateRlpPrefixLenDouble(uint64(len(b)), b[0]) + len(b)
+	return generateRlpPrefixLenDouble(len(b), b[0]) + len(b)
 }
 
 type RlpEncodedBytes []byte
@@ -48,13 +48,13 @@ func (b RlpEncodedBytes) RawBytes() []byte {
 }
 
 func (b RlpEncodedBytes) DoubleRLPLen() int {
-	return generateRlpPrefixLen(uint64(len(b))) + len(b)
+	return generateRlpPrefixLen(len(b)) + len(b)
 }
 
-func encodeBytesAsRlpToWriter(source []byte, w io.Writer, prefixGenFunc func([]byte, int, uint64) int, prefixBuf []byte) error {
+func encodeBytesAsRlpToWriter(source []byte, w io.Writer, prefixGenFunc func([]byte, int, int) int, prefixBuf []byte) error {
 	// > 1 byte, write a prefix or prefixes first
 	if len(source) > 1 || (len(source) == 1 && source[0] >= 0x80) {
-		prefixLen := prefixGenFunc(prefixBuf, 0, uint64(len(source)))
+		prefixLen := prefixGenFunc(prefixBuf, 0, len(source))
 
 		if _, err := w.Write(prefixBuf[:prefixLen]); err != nil {
 			return err
@@ -70,5 +70,5 @@ func EncodeByteArrayAsRlp(raw []byte, w io.Writer, prefixBuf []byte) (int, error
 	if err != nil {
 		return 0, err
 	}
-	return generateRlpPrefixLen(uint64(len(raw))) + len(raw), nil
+	return generateRlpPrefixLen(len(raw)) + len(raw), nil
 }
